Updated the Causes to share one folder and route through the index file.
