--- client/src/components/Private/Causes/style.css
@@ -1,4 +1,0 @@
-.styleButton {
-	background: linear-gradient(-135deg, #899fd4, #a389d4);
-	color: #ffffff !important;
-}
