+++ client/src/containers/Private/Dashboard/index.js
@@ -8,7 +8,7 @@ import Tab from "@material-ui/core/Tab";
 import Nav from "../../../components/Navigation";
 import Feed from "../../../components/Feed";
 import Gradient from "../../../components/Gradient";
-import Causes from "../../../components/Private/Causes";
+import Causes from "../../../components/Causes";
 import About from "../../../components/About";
 import Footer from "../../../components/Footer";
 import { TabPanel, a11yProps, useWindowDimensions } from "../../utils";
